Rename login scenes for clarity
