+++ src/main/resources/org/grupp2/sdpproject/database-login-scene.fxml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+
+<AnchorPane  prefHeight="424.0" prefWidth="600.0" stylesheets="@style.css" xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.grupp2.sdpproject.GUI.DatabaseLoginScene">
+   <children>
+      <VBox alignment="TOP_CENTER" prefHeight="424.0" prefWidth="600.0" styleClass="vbox">
+         <children>
+            <Label text="Skriv in inloggningsuppgifter till din MySQL databas." />
+            <Label text="Uppgifterna kommer sparas till fil och användas vid nästa start av programmet." />
+            <TextField fx:id="usernameField" maxWidth="200.0" promptText="Användarnamn" />
+            <TextField fx:id="passwordField" maxWidth="200.0" promptText="Lösenord" />
+            <TextField fx:id="ipField" maxWidth="200.0" promptText="IP-adress" text="localhost" />
+            <TextField fx:id="portField" maxWidth="200.0" promptText="Port" text="3306" />
+            <Label fx:id="messageLabel" styleClass="message-label" />
+            <HBox alignment="TOP_CENTER" prefHeight="23.0" prefWidth="580.0" styleClass="hbox">
+               <children>
+                  <Button maxWidth="90.0" mnemonicParsing="false" onAction="#handleLogin" text="Spara" />
+               </children>
+            </HBox>
+         </children>
+      </VBox>
+   </children>
+</AnchorPane>
