--- src/main/resources/org/grupp2/sdpproject/login2-scene.fxml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.geometry.*?>
-
-<AnchorPane fx:id="root" xmlns:fx="http://javafx.com/fxml"
-            fx:controller="org.grupp2.sdpproject.GUI.Login2Scene">
-    <VBox alignment="CENTER" spacing="10" layoutX="150" layoutY="100">
-        <Label text="Login" style="-fx-font-size: 18px;"/>
-        <TextField fx:id="emailField" promptText="Email"/>
-        <PasswordField fx:id="passwordField" promptText="Password"/>
-        <Button text="Login" onAction="#handleLogin"/>
-        <Hyperlink text="Register" onAction="#switchToRegister"/>
-        <Label fx:id="statusLabel" textFill="red"/>
-    </VBox>
-</AnchorPane>
